+++ app/src/main/java/com/kongjak/koreatechboard/ui/fragment/SimFragment.kt
@@ -1,4 +1,4 @@
-package com.kongjak.koreatechboard.fragment
+package com.kongjak.koreatechboard.ui.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,25 +7,27 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.kongjak.koreatechboard.R
 
-class EmcFragment : Fragment() {
+class SimFragment : Fragment() {
     lateinit var fragment: Fragment
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val rootView = inflater.inflate(R.layout.fragment_emc, container, false)
+        val rootView = inflater.inflate(R.layout.fragment_sim, container, false)
 
-        val emcNotice = BoardFragment()
-        val emcNoticeBundle = Bundle()
-        emcNoticeBundle.putString("board", "notice")
-        emcNoticeBundle.putString("site", "emc")
-        emcNotice.arguments = emcNoticeBundle
+        val simNotice = BoardFragment()
+        val simNoticeBundle = Bundle()
+        simNoticeBundle.putString("board", "notice")
+        simNoticeBundle.putString("site", "sim")
+        simNotice.arguments = simNoticeBundle
 
-        fragment = if (savedInstanceState == null) {
-            emcNotice
+        if (savedInstanceState == null) {
+            if (!this::fragment.isInitialized) {
+                fragment = simNotice
+            }
         } else {
-            parentFragmentManager.getFragment(savedInstanceState, "fragment")!!
+            fragment = parentFragmentManager.getFragment(savedInstanceState, "fragment")!!
         }
         loadFragment()
