+++ app/src/main/res/layout/activity_article.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.kongjak.koreatechboard.activity.ArticleActivity">
+    tools:context=".ui.activity.ArticleActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/app_bar_layout"
