+++ app/src/main/res/layout/fragment_board.xml
@@ -15,7 +15,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="com.kongjak.koreatechboard.fragment.BoardFragment">
+        tools:context=".ui.fragment.BoardFragment">
 
         <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
             android:id="@+id/swipe_refresh"
